pyproject.toml
[project]
name = "googleadk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk[mcp]>=1.17.0",
    "python-dotenv>=1.1.1",
]
